requirements.txt
torch
pillow
tqdm
transformers